// src/modules/files/enums/file-category.enum.ts
export enum FileCategory {
  PRODUCT_IMAGE = 'product_image',
  BRAND_LOGO = 'brand_logo',
  PROFILE_IMAGE = 'profile_image',
  PRESCRIPTION = 'prescription',
  MEDICAL_REPORT = 'medical_report',
  LABORATORY_REPORT = 'laboratory_report',
  ORGANIZATION_DOCUMENT = 'organization_document',
  INVOICE_DOCUMENT = 'invoice_document',
  SHIPPING_DOCUMENT = 'shipping_document',
  SUPPORT_ATTACHMENT = 'support_attachment',
  NOTIFICATION_ATTACHMENT = 'notification_attachment',
}