// src/modules/files/enums/resource-type.enum.ts
export enum ResourceType {
  BRAND_LOGO = 'brand_logo',
  USER_AVATAR = 'user_avatar',
  CUSTOMER_AVATAR = 'customer_avatar',
  PRODUCT_MEDIA = 'product_media',
  PRESCRIPTION_DOCUMENT = 'prescription_document',
  DIAGNOSTIC_REPORT = 'diagnostic_report',
  ORGANIZATION_LICENSE_DOCUMENT = 'organization_license_document',
  REGULATORY_REGISTRATION_DOCUMENT = 'regulatory_registration_document',
  INSURANCE_CLAIM_DOCUMENT = 'insurance_claim_document',
  SUPPORT_TICKET_ATTACHMENT = 'support_ticket_attachment',
  FINANCE_INVOICE_DOCUMENT = 'finance_invoice_document',
  FINANCE_CREDIT_NOTE_DOCUMENT = 'finance_credit_note_document',
  SHIPMENT_LABEL = 'shipment_label',
  DELIVERY_PROOF = 'delivery_proof',
  TELECONSULTATION_RECORDING = 'teleconsultation_recording',
  PAYMENT_RECONCILIATION_SOURCE = 'payment_reconciliation_source',
  SUPPLIER_LICENSE_DOCUMENT = 'supplier_license_document',
  SUPPLIER_INVOICE_DOCUMENT = 'supplier_invoice_document',
  NOTIFICATION_ATTACHMENT = 'notification_attachment',
  IN_APP_NOTIFICATION_IMAGE = 'in_app_notification_image',
}